pyproject.toml
[project]
name = "airbnb-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=1.4.1",
    "fastapi>=0.115.12",
    "mcp>=1.6.0",
    "openai>=1.76.0",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.40",
]
